Issue #2: Remove brace on package in model.scala
While scala allows one to use {} after a package name in order to define
some things in said package, it's uncommon (in my understanding) to use
the braces to encapsulate the whole file like it is in model.scala.

I've removed the braces and fixed the indentation in this commit.