Add initial OpenTypeFontTest
Using FlatSpec because it's a simple and easy to deal with test format
that doesn't throw weird exceptions like FunSuite if you have a mistake
in the class body.

I haven't looked over what the bytes mean, I'm just basing the bytes
themselves off the main method output in the test class.