Add OTFByteAllocator#allocate tests
I'm still not entirely sure about everything going on in this, but
figure that the byte allocators seem to have the most logic and it might
make sense to start adding tests around said logic.

I didn't add any tests for the allocate method with the data parameter
because I'm not sure I understand the allocMap's purpose in the file
yet.