Refactor mapping in SegmentedCoverageEncodingFormat
Addressing Issue #12

Besides removing the flatMap + map + map + toMap, I've made the
retrieval of the glyph IDs more safe. Before it was range.drop(i).head
which was an unsafe operation since .head can produce a NoSuchElement
exception. I was reading the documentation linked to on the issue and
noticed that it said (near the top) that if there is no glyph
corresponding with the codepoint, then it should map to 0. So, the drop
+ head combination becomes `.lift(i)` which gives you an Option only
filled with a value if the glyph was valid, then this option is folded
into a UInt or to the default emptyGlyph